Bike Pump Shooting Works
I was able to make the player shoot the pump and should it miss, it would return back to the dig state. Instead of working on the attack state. It would be a better idea to make a basic enemy first and then implement the attack state. I was gonna work on the level but I found out the sprite's don't align properly. I will either get a different set of sprites for that or edit it just enough for it to work.