Completed the main Objective
The main objective of this game was to create a destructible environment which I was finally able to complete. It took a while for me to get it working properly. I also made sure it works properly by making the player to the middle of the tile to accurately hit the tile normal. I gonna consider this project completed as I want to move on to the project and learned what I had to.